Add files via upload
This project consists of three separate functions designed to analyze customer feedback on a newly launched product. It also calculates the number of customers whose purchases exceed $100 or more. The project aims to provide insights into product satisfaction and customer spending habits.

Customer feedback analysis

The number_of_feedback function takes a list of numerical scores (ranging from 1 to 10) as input, representing customer feedback for a product. It categorizes feedback into negative, neutral, and positive based on the score ranges provided in the code. The function then prints the counts of each category.

Fake ID detection

The id-valid  function assesses the validity of a list of IDs by comparing them to a list of accepted IDs. It calculates the number and percentage of fake IDs in the given list. This can be useful for verification purposes.

Purchase Sum Calculator .

The purchases_100 function takes a list of lists (sales data) where each inner list contains the prices of items purchased by each customer. It calculates the total purchase value for each customer and counts the number of customers whose purchases amount to $100 or more.